Add supported services to the initial state
